Refactor routing
